+++ src/lint/utils/math.ts
@@ -1,0 +1,8 @@
+
+export const max = (nums: number[]) => {
+  let _max = -Infinity;
+  nums.forEach(num => {
+    _max = Math.max(num, _max);
+  })
+  return _max;
+}
